+++ Makefile
@@ -3,7 +3,7 @@ CC = clang
 RM = rm -rf
 SRCSDIR = srcs
 OBJSDIR = objs
-SRCS = main.c utils/libfunc.c utils/libfunc_two.c exec/print_exec_error.c exec/miniparser.c exec/parser_assets.c exec/execution.c exec/freer.c exec/redir_parse.c exec/new_table.c exec/child_processes.c exec/exec_assets.c exec/path_handler_p2.c exec/exec_no_pipe.c\
+SRCS = main.c utils/libfunc.c utils/libfunc_two.c exec/print_exec_error.c exec/miniparser.c exec/parser_assets.c exec/execution.c exec/freer.c exec/redir_parse.c exec/new_table.c exec/child_processes.c exec/exec_assets.c exec/path_handler_p2.c exec/exec_no_pipe.c exec/cmds_parse.c exec/sig.c exec/lim_exec.c exec/fds_closer.c exec/parent_delim.c\
 parsing/check_parsing_errors.c parsing/ft_init.c parsing/ft_return.c parsing/handle_else.c parsing/ft_itoa.c builtin/ft_env.c builtin/ft_exit.c builtin/ft_export_two.c\
 parsing/lexer.c parsing/parser_init.c parsing/token_list_init.c signal_init.c signal_errors.c builtin/ft_export.c\
 parsing/handle_quote.c utils/utils_2.c parsing/is_env.c parsing/parser.c utils/utils.c exec/path_handler.c exec/builtin.c builtin/ft_cd.c\
