+++ srcs/exec/exec_assets.c
@@ -6,13 +6,11 @@
 /*   By: ybendavi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/29 18:42:15 by ybendavi          #+#    #+#             */
-/*   Updated: 2022/07/07 20:51:02 by ybendavi         ###   ########.fr       */
+/*   Updated: 2022/07/08 15:05:32 by ybendavi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-extern int	g_sig;
 
 char	**path_selector(char **paths)
 {	
@@ -79,53 +77,4 @@ int	set_forks(t_env	*envs)
 		tmp = tmp->next;
 	}
 	return (0);
-}
-
-void	parent_process(t_cmds *cmds, int status, t_env *envs)
-{	
-	if (cmds->in != 0 && cmds->in != -1 && cmds->in != -3)
-	{
-		close(cmds->in);
-		cmds->in = -3;
-	}
-	if (cmds->prev && cmds->prev->out != 1 && cmds->prev->out != -1
-		&& cmds->prev->out != -3)
-	{
-		close(cmds->prev->out);
-		cmds->prev->out = -3;
-	}
-	if (cmds->delim)
-	{
-		if (cmds->lim[0] != -1 && cmds->lim[0] != -3)
-		{
-			close(cmds->lim[0]);
-			cmds->lim[0] = -3;
-		}
-		if (cmds->lim[1] != -1 && cmds->lim[1] != -3)
-		{
-			close(cmds->lim[1]);
-			cmds->lim[1] = -3;
-		}
-		envs->sig_i.sa_handler = &kill_int;
-		sigaction(SIGINT, &(envs->sig_i), NULL);
-		if (g_sig == 42)
-			kill(cmds->fork, SIGINT);
-		waitpid(cmds->fork, &status, 0);
-		if (WIFEXITED(status) != 0)
-		{
-			printf("status:%d\n", WEXITSTATUS(status));
-			//if (WEXITSTATUS(*status) == 130)
-			//{
-		//		g_sig = 42;
-		//		printf("global:%d\n", g_sig);
-		//	}
-		}
-		//if (!WIFSIGNALED(*status))
-		//{
-			//if (WTERMSIG(*status) == SIGINT)
-			//	g_sig = 42;
-		//}
-		envs->sig_i.sa_handler = &handler_sig;
-		sigaction(SIGINT, &(envs->sig_i), NULL);
-	}
 }
