+++ srcs/exec/parent_delim.c
@@ -1,0 +1,52 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   parent_delim.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ybendavi <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/07/08 15:02:25 by ybendavi          #+#    #+#             */
+/*   Updated: 2022/07/08 15:04:57 by ybendavi         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell"
+
+void	delim_parent(t_cmds *cmds, int status, t_env *envs)
+{	
+	if (cmds->lim[0] != -1 && cmds->lim[0] != -3)
+	{
+		close(cmds->lim[0]);
+		cmds->lim[0] = -3;
+	}
+	if (cmds->lim[1] != -1 && cmds->lim[1] != -3)
+	{
+		close(cmds->lim[1]);
+		cmds->lim[1] = -3;
+	}
+	envs->sig_i.sa_handler = &kill_int;
+	sigaction(SIGINT, &(envs->sig_i), NULL);
+	if (check_global == 1)
+		kill(cmds->fork, SIGINT);
+	waitpid(cmds->fork, &status, 0);
+	envs->sig_i.sa_handler = &handler_sig;
+	sigaction(SIGINT, &(envs->sig_i), NULL);
+}
+
+void	parent_process(t_cmds *cmds, int status, t_env *envs)
+{
+	if (cmds->in != 0 && cmds->in != -1 && cmds->in != -3)
+	{
+		close(cmds->in);
+		cmds->in = -3;
+	}
+	if (cmds->prev && cmds->prev->out != 1 && cmds->prev->out != -1
+		&& cmds->prev->out != -3)
+	{
+		close(cmds->prev->out);
+		cmds->prev->out = -3;
+	}
+	if (cmds->delim)
+		delim_parent(cmds, status, envs);
+}
+
